app/requirements.txt
flask
PyMySQL
requests
mysqlclient
sqlalchemy
mysql-connector-python
pyjwt
flask_cors
